Car-Leasing-Project v0.3
add management poossibilities to cars, need to do it for clients. Little bugs were fixed

+++ src/pl/pjatk/Car/CarManagement.java
@@ -2,7 +2,6 @@ package pl.pjatk.Car;
 
 import pl.pjatk.Management.Management;
 
-import java.util.Scanner;
 
 public class CarManagement {
     public void startManagement(){
@@ -10,16 +9,31 @@ public class CarManagement {
         System.out.println();
         System.out.println("1. Dodaj auto");
         System.out.println("2. Wyświetl auta");
-        System.out.println("3. Zmień dane auta");
+        System.out.println("3. Zarządzaj autem");
         System.out.println("4. Usuń auto");
         System.out.println("0. Cofnij.");
         System.out.println("Wybierz operację: ");
 
-            switch (Management.choosing(4)) {
+            switch (Management.choosingSwitchInput(4)) {
                 case 1 -> addCar();
-                case 2 -> showCars();
-                case 3 -> changeCar();
+                case 2 -> showCars()
+                case 3 -> managementCar();
                 case 4 -> deleteCar();
+            }
+            break;
+        }
+    }
+
+    public void managementCar(){
+        while(true) {
+            System.out.println();
+            System.out.println("1. Zaktualizuj cenę auta");
+            System.out.println("2. Zmień dostępność auta");
+            System.out.println("Wybierz operację: ");
+
+            switch (Management.choosingSwitchInput(2)) {
+                case 1 -> updatePriceCar();
+                case 2 -> changeAvailabiltyCar();
             }
             break;
         }
@@ -28,19 +42,16 @@ public class CarManagement {
     public void addCar(){
         CarDataSource carDataSource = new CarDataSource();
 
-        Scanner scanner = new Scanner(System.in);
         System.out.println("Podaj markę: ");
-        String mark = scanner.nextLine();
+        String mark = Management.choosingStringInput();
         System.out.println("Podaj model: ");
-        String model = scanner.nextLine();
+        String model = Management.choosingStringInput();
         System.out.println("Podaj rocznik: ");
-        int yearProduced = scanner.nextInt();
-        scanner.nextLine();
+        int yearProduced = Management.choosingIntInput();
         System.out.println("Podaj cenę: ");
-        double price = scanner.nextDouble();
-        scanner.nextLine();
+        double price = Management.choosingDoubleInput();
         System.out.println("Podaj numer rejestracyjny: ");
-        String registrationNumber = scanner.nextLine();
+        String registrationNumber = Management.choosingStringInput();
 
         int idSpecificCar = carDataSource.getNumberOfSpecificCars()+1;
         int idModel = carDataSource.selectModelId(mark, model);
@@ -67,16 +78,39 @@ public class CarManagement {
         carDataSource.selectCars();
     }
 
-    public void changeCar(){
+    public void updatePriceCar(){
+        CarDataSource carDataSource = new CarDataSource();
+        System.out.println("Wpisz numer rejestracyjny auta, którego cenę chcesz zmienić: ");
+        String registrationNumber = Management.choosingStringInput();
+        if(carDataSource.isExistCar(registrationNumber)){
+            System.out.println("Podaj nową cenę: ");
+            double newPrice = Management.choosingDoubleInput();
+            carDataSource.updatePriceCar(registrationNumber,newPrice);
+        } else {
+            System.out.println("Nie mamy takiego samochodu");
+        }
+    }
 
+    public void changeAvailabiltyCar(){
+        CarDataSource carDataSource = new CarDataSource();
+        System.out.println("Wpisz numer rejestracyjny auta, którego dostępność chcesz zmienić: ");
+        String registrationNumber = Management.choosingStringInput();
+        if(carDataSource.isExistCar(registrationNumber)){
+            carDataSource.updateAvailability(registrationNumber);
+        } else {
+            System.out.println("Nie mamy takiego samochodu");
+        }
     }
 
     public void deleteCar(){
-        Scanner scanner = new Scanner(System.in);
         CarDataSource carDataSource = new CarDataSource();
         carDataSource.selectCars();
         System.out.println("Wpisz numer rejestracyjny auta, które chcesz usunąć: ");
-        String registrationNumber = scanner.nextLine();
-        carDataSource.deleteCar(registrationNumber);
+        String registrationNumber = Management.choosingStringInput();
+        if(carDataSource.isExistCar(registrationNumber)){
+            carDataSource.deleteCar(registrationNumber);
+        } else {
+            System.out.println("Nie mamy takiego samochodu");
+        }
     }
 }
